+++ beacon-chain/core/altair/transition_no_verify_sig_test.go
@@ -1,0 +1,229 @@
+package altair_test
+
+import (
+	"context"
+	"testing"
+
+	types "github.com/prysmaticlabs/eth2-types"
+	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+	"github.com/prysmaticlabs/go-bitfield"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/altair"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+	p2pType "github.com/prysmaticlabs/prysm/beacon-chain/p2p/types"
+	iface "github.com/prysmaticlabs/prysm/beacon-chain/state/interface"
+	"github.com/prysmaticlabs/prysm/shared/attestationutil"
+	"github.com/prysmaticlabs/prysm/shared/bls"
+	"github.com/prysmaticlabs/prysm/shared/bytesutil"
+	"github.com/prysmaticlabs/prysm/shared/params"
+	"github.com/prysmaticlabs/prysm/shared/testutil"
+	testutilAltair "github.com/prysmaticlabs/prysm/shared/testutil/altair"
+	"github.com/prysmaticlabs/prysm/shared/testutil/require"
+)
+
+func TestProcessBlockNoVerify_PassesProcessingConditions(t *testing.T) {
+	beaconState, block, _, _, _ := createFullBlockWithOperations(t)
+	set, _, err := altair.ProcessBlockNoVerifyAnySig(context.Background(), beaconState, block)
+	require.NoError(t, err)
+	// Test Signature set verifies.
+	verified, err := set.Verify()
+	require.NoError(t, err)
+	require.Equal(t, true, verified, "Could not verify signature set")
+}
+
+func createFullBlockWithOperations(t *testing.T) (iface.BeaconStateAltair,
+	*ethpb.SignedBeaconBlockAltair, []*ethpb.Attestation, []*ethpb.ProposerSlashing, []*ethpb.SignedVoluntaryExit) {
+	beaconState, privKeys := testutilAltair.DeterministicGenesisStateAltair(t, 32)
+	genesisBlock := blocks.NewGenesisBlock([]byte{})
+	bodyRoot, err := genesisBlock.Block.HashTreeRoot()
+	require.NoError(t, err)
+	err = beaconState.SetLatestBlockHeader(&ethpb.BeaconBlockHeader{
+		Slot:       genesisBlock.Block.Slot,
+		ParentRoot: genesisBlock.Block.ParentRoot,
+		StateRoot:  params.BeaconConfig().ZeroHash[:],
+		BodyRoot:   bodyRoot[:],
+	})
+	require.NoError(t, err)
+	err = beaconState.SetSlashings(make([]uint64, params.BeaconConfig().EpochsPerSlashingsVector))
+	require.NoError(t, err)
+	cp := beaconState.CurrentJustifiedCheckpoint()
+	mockRoot := [32]byte{}
+	copy(mockRoot[:], "hello-world")
+	cp.Root = mockRoot[:]
+	require.NoError(t, beaconState.SetCurrentJustifiedCheckpoint(cp))
+
+	proposerSlashIdx := types.ValidatorIndex(3)
+	slotsPerEpoch := params.BeaconConfig().SlotsPerEpoch
+	err = beaconState.SetSlot(slotsPerEpoch.Mul(uint64(params.BeaconConfig().ShardCommitteePeriod)) + params.BeaconConfig().MinAttestationInclusionDelay)
+	require.NoError(t, err)
+
+	currentEpoch := helpers.CurrentEpoch(beaconState)
+	header1 := testutil.HydrateSignedBeaconHeader(&ethpb.SignedBeaconBlockHeader{
+		Header: &ethpb.BeaconBlockHeader{
+			ProposerIndex: proposerSlashIdx,
+			Slot:          1,
+			StateRoot:     bytesutil.PadTo([]byte("A"), 32),
+		},
+	})
+	header1.Signature, err = helpers.ComputeDomainAndSign(beaconState, currentEpoch, header1.Header, params.BeaconConfig().DomainBeaconProposer, privKeys[proposerSlashIdx])
+	require.NoError(t, err)
+
+	header2 := testutil.HydrateSignedBeaconHeader(&ethpb.SignedBeaconBlockHeader{
+		Header: &ethpb.BeaconBlockHeader{
+			ProposerIndex: proposerSlashIdx,
+			Slot:          1,
+			StateRoot:     bytesutil.PadTo([]byte("B"), 32),
+		},
+	})
+	header2.Signature, err = helpers.ComputeDomainAndSign(beaconState, helpers.CurrentEpoch(beaconState), header2.Header, params.BeaconConfig().DomainBeaconProposer, privKeys[proposerSlashIdx])
+	require.NoError(t, err)
+
+	proposerSlashings := []*ethpb.ProposerSlashing{
+		{
+			Header_1: header1,
+			Header_2: header2,
+		},
+	}
+	validators := beaconState.Validators()
+	validators[proposerSlashIdx].PublicKey = privKeys[proposerSlashIdx].PublicKey().Marshal()
+	require.NoError(t, beaconState.SetValidators(validators))
+
+	mockRoot2 := [32]byte{'A'}
+	att1 := testutil.HydrateIndexedAttestation(&ethpb.IndexedAttestation{
+		Data: &ethpb.AttestationData{
+			Source: &ethpb.Checkpoint{Epoch: 0, Root: mockRoot2[:]},
+		},
+		AttestingIndices: []uint64{0, 1},
+	})
+	domain, err := helpers.Domain(beaconState.Fork(), currentEpoch, params.BeaconConfig().DomainBeaconAttester, beaconState.GenesisValidatorRoot())
+	require.NoError(t, err)
+	hashTreeRoot, err := helpers.ComputeSigningRoot(att1.Data, domain)
+	require.NoError(t, err)
+	sig0 := privKeys[0].Sign(hashTreeRoot[:])
+	sig1 := privKeys[1].Sign(hashTreeRoot[:])
+	aggregateSig := bls.AggregateSignatures([]bls.Signature{sig0, sig1})
+	att1.Signature = aggregateSig.Marshal()
+
+	mockRoot3 := [32]byte{'B'}
+	att2 := testutil.HydrateIndexedAttestation(&ethpb.IndexedAttestation{
+		Data: &ethpb.AttestationData{
+			Source: &ethpb.Checkpoint{Epoch: 0, Root: mockRoot3[:]},
+			Target: &ethpb.Checkpoint{Epoch: 0, Root: make([]byte, 32)},
+		},
+		AttestingIndices: []uint64{0, 1},
+	})
+
+	hashTreeRoot, err = helpers.ComputeSigningRoot(att2.Data, domain)
+	require.NoError(t, err)
+	sig0 = privKeys[0].Sign(hashTreeRoot[:])
+	sig1 = privKeys[1].Sign(hashTreeRoot[:])
+	aggregateSig = bls.AggregateSignatures([]bls.Signature{sig0, sig1})
+	att2.Signature = aggregateSig.Marshal()
+
+	attesterSlashings := []*ethpb.AttesterSlashing{
+		{
+			Attestation_1: att1,
+			Attestation_2: att2,
+		},
+	}
+
+	var blockRoots [][]byte
+	for i := uint64(0); i < uint64(params.BeaconConfig().SlotsPerHistoricalRoot); i++ {
+		blockRoots = append(blockRoots, []byte{byte(i)})
+	}
+	require.NoError(t, beaconState.SetBlockRoots(blockRoots))
+
+	aggBits := bitfield.NewBitlist(1)
+	aggBits.SetBitAt(0, true)
+	blockAtt := testutil.HydrateAttestation(&ethpb.Attestation{
+		Data: &ethpb.AttestationData{
+			Slot:   beaconState.Slot(),
+			Target: &ethpb.Checkpoint{Epoch: helpers.CurrentEpoch(beaconState)},
+			Source: &ethpb.Checkpoint{Root: mockRoot[:]}},
+		AggregationBits: aggBits,
+	})
+
+	committee, err := helpers.BeaconCommitteeFromState(beaconState, blockAtt.Data.Slot, blockAtt.Data.CommitteeIndex)
+	require.NoError(t, err)
+	attestingIndices, err := attestationutil.AttestingIndices(blockAtt.AggregationBits, committee)
+	require.NoError(t, err)
+	require.NoError(t, err)
+	hashTreeRoot, err = helpers.ComputeSigningRoot(blockAtt.Data, domain)
+	require.NoError(t, err)
+	sigs := make([]bls.Signature, len(attestingIndices))
+	for i, indice := range attestingIndices {
+		sig := privKeys[indice].Sign(hashTreeRoot[:])
+		sigs[i] = sig
+	}
+	blockAtt.Signature = bls.AggregateSignatures(sigs).Marshal()
+
+	exit := &ethpb.SignedVoluntaryExit{
+		Exit: &ethpb.VoluntaryExit{
+			ValidatorIndex: 10,
+			Epoch:          0,
+		},
+	}
+	exit.Signature, err = helpers.ComputeDomainAndSign(beaconState, currentEpoch, exit.Exit, params.BeaconConfig().DomainVoluntaryExit, privKeys[exit.Exit.ValidatorIndex])
+	require.NoError(t, err)
+
+	header := beaconState.LatestBlockHeader()
+	prevStateRoot, err := beaconState.HashTreeRoot(context.Background())
+	require.NoError(t, err)
+	header.StateRoot = prevStateRoot[:]
+	require.NoError(t, beaconState.SetLatestBlockHeader(header))
+	parentRoot, err := beaconState.LatestBlockHeader().HashTreeRoot()
+	require.NoError(t, err)
+	copied := beaconState.Copy()
+	require.NoError(t, copied.SetSlot(beaconState.Slot()+1))
+	randaoReveal, err := testutil.RandaoReveal(copied, currentEpoch, privKeys)
+	require.NoError(t, err)
+	proposerIndex, err := helpers.BeaconProposerIndex(copied)
+	require.NoError(t, err)
+
+	syncBits := make([]byte, 1)
+	for i := range syncBits {
+		syncBits[i] = 0xff
+	}
+	indices, err := altair.SyncCommitteeIndices(beaconState, helpers.CurrentEpoch(beaconState))
+	require.NoError(t, err)
+	ps := helpers.PrevSlot(beaconState.Slot())
+	pbr, err := helpers.BlockRootAtSlot(beaconState, ps)
+	require.NoError(t, err)
+	syncSigs := make([]bls.Signature, len(indices))
+	for i, indice := range indices {
+		b := p2pType.SSZBytes(pbr)
+		sb, err := helpers.ComputeDomainAndSign(beaconState, helpers.CurrentEpoch(beaconState), &b, params.BeaconConfig().DomainSyncCommittee, privKeys[indice])
+		require.NoError(t, err)
+		sig, err := bls.SignatureFromBytes(sb)
+		require.NoError(t, err)
+		syncSigs[i] = sig
+	}
+	aggregatedSig := bls.AggregateSignatures(syncSigs).Marshal()
+	syncAggregate := &ethpb.SyncAggregate{
+		SyncCommitteeBits:      syncBits,
+		SyncCommitteeSignature: aggregatedSig,
+	}
+
+	block := testutil.HydrateSignedBeaconBlockAltair(&ethpb.SignedBeaconBlockAltair{
+		Block: &ethpb.BeaconBlockAltair{
+			ParentRoot:    parentRoot[:],
+			Slot:          beaconState.Slot() + 1,
+			ProposerIndex: proposerIndex,
+			Body: &ethpb.BeaconBlockBodyAltair{
+				RandaoReveal:      randaoReveal,
+				ProposerSlashings: proposerSlashings,
+				AttesterSlashings: attesterSlashings,
+				Attestations:      []*ethpb.Attestation{blockAtt},
+				VoluntaryExits:    []*ethpb.SignedVoluntaryExit{exit},
+				SyncAggregate:     syncAggregate,
+			},
+		},
+	})
+
+	sig, err := testutil.BlockSignatureAltair(beaconState, block.Block, privKeys)
+	require.NoError(t, err)
+	block.Signature = sig.Marshal()
+
+	require.NoError(t, beaconState.SetSlot(block.Block.Slot))
+	return beaconState, block, []*ethpb.Attestation{blockAtt}, proposerSlashings, []*ethpb.SignedVoluntaryExit{exit}
+}
